fix: karty-widma zarchiwizowanych planow znikaja z karty dnia
Wpisy MOVED/SKIPPED zarchiwizowanego planu wisialy w kalendarzu na
zawsze: flaga ScheduleEntryRef.archived byla ustawiana wylacznie dla
PLANNED, wiec sweep martwych wpisow (archivedPlanDeadEntryIds) nigdy
ich nie widzial, a shadowedEntryIds kluczuje po (planId, data) i pary
"widmo starego planu obok treningu nowego" z definicji nie lapie.

- archived w obu budowniczych refow (ScheduleViewModel,
  PlanEditorViewModel) dla kazdego statusu poza DONE;
- archivedPlanDeadEntryIds kasuje przyszle PLANNED/MOVED/SKIPPED
  zarchiwizowanych planow, DONE nigdy;
- nowe archivedPlanGhostEntryIds ukrywa MOVED/SKIPPED zarchiwizowanych
  planow w karcie dnia od razu, niezaleznie od daty (baza trzyma
  przeszlosc jako audit-trail, oczy jej nie ogladaja);
- snackbar "Trening odwolany." dostaje SnackbarDuration.Long - przy
  Short nie dalo sie zdazyc z "Cofnij".

+++ app/src/main/java/com/stronk/ui/schedule/ScheduleScreen.kt
@@ -117,13 +117,17 @@ fun ScheduleScreen(
     // ta akcja jest jedyną siatką bezpieczeństwa. `dismiss()` zamyka ewentualny
     // wciąż widoczny poprzedni snackbar odwołania, żeby „Cofnij" nigdy nie
     // trafiało w nieaktualny (już zamknięty pod spodem) wpis.
+    // [SnackbarDuration.Long], nie Short: przy Short snackbar znikał, zanim
+    // dało się trafić w „Cofnij" (gate: 2 z 3 prób nie zdążyły) — okno na
+    // wycofanie przypadkowego tapu musi być realne, inaczej cała siatka
+    // bezpieczeństwa jest dekoracją.
     LaunchedEffect(cancelEvent) {
         val event = cancelEvent ?: return@LaunchedEffect
         snackbarHostState.currentSnackbarData?.dismiss()
         val result = snackbarHostState.showSnackbar(
             message = ScheduleTexts.WORKOUT_CANCELLED,
             actionLabel = ScheduleTexts.UNDO_CANCEL,
-            duration = SnackbarDuration.Short,
+            duration = SnackbarDuration.Long,
         )
         if (result == SnackbarResult.ActionPerformed) {
             viewModel.onRestoreEntry(event.entryId)
